requirements.txt
Django==5.2.9
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-celery-beat==2.8.1
celery==5.6.1
redis==7.1.0
python-decouple==3.8
Pillow==12.0.0
qrcode==8.2

# Opcional para producción
# psycopg2-binary==2.9.9  # PostgreSQL
# gunicorn==21.2.0  # Server WSGI
# whitenoise==6.6.0  # Static files
